requirements.txt
numpy
scikit-learn
opencv-python
pillow
tqdm
einops==0.7.0
torch==2.1.2
torchvision==0.16.2
timm==0.9.12
fvcore